pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spectroscopy-auc"
version = "0.1.0"
description = "A pipeline for spectroscopy Area Under the Curve (AUC) extraction and visualisation"
authors = [
    {name = "Hugo DESSAUD", email = "hdessaud1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT License"}
dependencies = [
    "numpy",
    "matplotlib",
    "pillow",
    "scipy",
    "streamlit",
    "pandas"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "isort",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88
